+++ CHANGELOG.md
@@ -89,3 +89,4 @@
 * Define `Adjunction` type family
 * Define `Bindable` type family
 * Define `Traversable` type family
+* Define `|-/->` operator for `Bindable` functors